Implement gl calls (see on the screen!), convert doubles to floats

+++ cpp/math/Frustum.cpp
@@ -5,22 +5,22 @@ Frustum::Frustum() {}
 Frustum* Frustum::setFromMatrix(
 	Matrix4* m
 ) {
-	double me0 = m->elements[0];
-	double me1 = m->elements[1];
-	double me2 = m->elements[2];
-	double me3 = m->elements[3];
-	double me4 = m->elements[4];
-	double me5 = m->elements[5];
-	double me6 = m->elements[6];
-	double me7 = m->elements[7];
-	double me8 = m->elements[8];
-	double me9 = m->elements[9];
-	double me10 = m->elements[10];
-	double me11 = m->elements[11];
-	double me12 = m->elements[12];
-	double me13 = m->elements[13];
-	double me14 = m->elements[14];
-	double me15 = m->elements[15];
+	float me0 = m->elements[0];
+	float me1 = m->elements[1];
+	float me2 = m->elements[2];
+	float me3 = m->elements[3];
+	float me4 = m->elements[4];
+	float me5 = m->elements[5];
+	float me6 = m->elements[6];
+	float me7 = m->elements[7];
+	float me8 = m->elements[8];
+	float me9 = m->elements[9];
+	float me10 = m->elements[10];
+	float me11 = m->elements[11];
+	float me12 = m->elements[12];
+	float me13 = m->elements[13];
+	float me14 = m->elements[14];
+	float me15 = m->elements[15];
 
 	this->planes[0].setComponents(me3 - me0, me7 - me4, me11 - me8, me15 - me12)->normalize();
 	this->planes[1].setComponents(me3 + me0, me7 + me4, me11 + me8, me15 + me12)->normalize();
